requirements.txt
fastapi
uvicorn
pandas
xgboost
python-dotenv
psycopg2-binary
scikit-learn
sqlalchemy
openai
google-generativeai